Решена задача на генерацию правильных скобочных последовательностей

+++ Ya_Practicum/Q.cpp
@@ -27,7 +27,7 @@ bool isOnOneLine(const vector<Point>& points) {
     long long del_x, del_y, cur_y, cur_x; // исходное k = del_y/del_x
     del_x = points[i].x -  points[0].x;
     del_y = points[i].y - points[0].y;
-    int x0 = points[0].x, y0 = points[0].y;
+    long long  x0 = points[0].x, y0 = points[0].y;
     i++; // Теперь i - первая из непроверенных точек. Но она может быть и концом списка
     for( ;i<points.size(); i++){
         if(points[i].x == x0 && points[i].y == y0){
